+++ calc.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
-int add(int a, int b);
+long long add(int a, int b);
 
 int main() {
 	int a, b;
 	scanf("%d %d", &a, &b);
-	printf("%d + %d = %d\n", a, b, add(a, b));
+	printf("%d + %d = %lld\n", a, b, add(a, b));
 }
